Bcc.Api.Address

Description

Bcc addresses: payment and stake addresses.

Payment addresses

Constructing and inspecting normal payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Cole addresses, which use the type tag ColeAddr; and

Sophie addresses, which use the type tag SophieAddr. Notably, Sophie addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Cole addresses are the only choice in the Cole era, the Sophie era and all subsequent eras support both Cole and Sophie addresses. The Address type param only says the type of the address (either Cole or Sophie). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ColeAddress ∷ Address → Address ColeAddr
Cole addresses were the only supported address type in the original Cole era.

SophieAddress ∷ Network → PaymentCredential StandardCrypto → StakeReference StandardCrypto → Address SophieAddr
Sophie addresses allow delegation. Sophie addresses were introduced in Sophie era and are thus supported from the Sophie era onwards

Cole addresses

data ColeAddr

A type used as a tag to distinguish Cole addresses.

makeColeAddress ∷ NetworkId → VerificationKey ColeKey → Address ColeAddr

Sophie addresses

data SophieAddr

A type used as a tag to distinguish Sophie addresses.

makeSophieAddress ∷ NetworkId → PaymentCredential → StakeAddressReference → Address SophieAddr

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

newtype StakeAddressPointer

Constructors

StakeAddressPointer

Fields

unStakeAddressPointer ∷ Ptr

Addresses in any era

data AddressAny

Either a Cole address or a Sophie address.

Sometimes we need to be able to work with either of the two types of address (Cole or Sophie addresses), but without reference to an era in which the address will be used. This type serves that purpose.

Constructors

AddressCole !(Address ColeAddr)

AddressSophie !(Address SophieAddr)

Addresses in specific eras

data AddressInEra era where

An Address that can be used in a particular ledger era.

All current ledger eras support Cole addresses. Sophie addresses are supported in the SophieEra and later eras.

Constructors

AddressInEra ∷ AddressTypeInEra addrtype era → Address addrtype → AddressInEra era

data AddressTypeInEra addrtype era where

Constructors

ColeAddressInAnyEra ∷ AddressTypeInEra ColeAddr era

SophieAddressInEra ∷ SophieBasedEra era → AddressTypeInEra SophieAddr era

coleAddressInEra ∷ Address ColeAddr → AddressInEra era

sophieAddressInEra ∷ IsSophieBasedEra era ⇒ Address SophieAddr → AddressInEra era

anyAddressInSophieBasedEra ∷ IsSophieBasedEra era ⇒ AddressAny → AddressInEra era

anyAddressInEra ∷ BccEra era → AddressAny → Maybe (AddressInEra era)

toAddressAny ∷ Address addr → AddressAny

makeColeAddressInEra ∷ NetworkId → VerificationKey ColeKey → AddressInEra era

makeSophieAddressInEra ∷ IsSophieBasedEra era ⇒ NetworkId → PaymentCredential → StakeAddressReference → AddressInEra era

Stake addresses

Constructing and inspecting stake addresses

data StakeAddress where

Constructors

StakeAddress ∷ Network → StakeCredential StandardCrypto → StakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

makeStakeAddress ∷ NetworkId → StakeCredential → StakeAddress

data StakeKey

data StakeExtendedKey

Sophie-era stake keys using extended ed25519 cryptographic keys.

They can be used for Sophie stake addresses and witnessing transactions that use stake addresses.

These extended keys are used by HD wallets. So this type provides interoperability with HD wallets. The ITN CLI also supported this key type.

The extended verification keys can be converted (via castVerificationKey) to ordinary keys (i.e. VerificationKey StakeKey) but this is not the case for the signing keys. The signing keys can be used to witness transactions directly, with verification via their non-extended verification key (VerificationKey StakeKey).

This is a type level tag, used with other interfaces like Key.

Internal conversion functions

toSophieAddr ∷ AddressInEra era → Addr StandardCrypto

toSophieStakeAddr ∷ StakeAddress → RewardAcnt StandardCrypto

toSophieStakeCredential ∷ StakeCredential → StakeCredential StandardCrypto

fromSophieAddr ∷ IsSophieBasedEra era ⇒ Addr StandardCrypto → AddressInEra era

fromSophiePaymentCredential ∷ PaymentCredential StandardCrypto → PaymentCredential

fromSophieStakeAddr ∷ RewardAcnt StandardCrypto → StakeAddress

fromSophieStakeCredential ∷ StakeCredential StandardCrypto → StakeCredential

fromSophieStakeReference ∷ StakeReference StandardCrypto → StakeAddressReference

Serialising addresses

class HasTypeProxy addr ⇒ SerialiseAddress addr where

Address serialisation uses different serialisation formats for different kinds of addresses, so it needs its own class.

In particular, Cole addresses are typically formatted in base 58, while Sophie addresses (payment and stake) are formatted using Bech32.

Methods

serialiseAddress ∷ addr → Text

deserialiseAddress ∷ AsType addr → Text → Maybe addr

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Helpers

isKeyAddress ∷ AddressInEra era → Bool

Is the UTxO at the address only spendable via a key witness.