Bcc.Api.Block

Description

Blocks in the blockchain

Blocks in the context of an era

data Block era where

A blockchain block in a particular Bcc era.

Constructors

ColeBlock ∷ ColeBlock → Block ColeEra

SophieBlock ∷ SophieBasedEra era → SophieBlock (SophieLedgerEra era) → Block era

Bundled Patterns

pattern Block ∷ BlockHeader → [Tx era] → Block era
A block consists of a header and a body containing transactions.

data BlockHeader

Constructors

BlockHeader !SlotNo !(Hash BlockHeader) !BlockNo

Blocks in the context of a consensus mode

data BlockInMode mode where

A Block in one of the eras supported by a given protocol mode.

For multi-era modes such as the BccMode this type is a sum of the different block types for all the eras. It is used in the ChainSync protocol.

Constructors

BlockInMode ∷ Block era → EraInMode era mode → BlockInMode mode

fromConsensusBlock ∷ ConsensusBlockForMode mode ~ block ⇒ ConsensusMode mode → block → BlockInMode mode

Points on the chain

data ChainPoint

Constructors

ChainPointAtGenesis

ChainPoint !SlotNo !(Hash BlockHeader)

newtype SlotNo

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

toConsensusPoint ∷ ∀ ledgerera. SophieBasedEra ledgerera ⇒ ChainPoint → Point (SophieBlock ledgerera)

Convert a Point for single Sophie-era block type

fromConsensusPoint ∷ ∀ ledgerera. SophieBasedEra ledgerera ⇒ Point (SophieBlock ledgerera) → ChainPoint

Convert a Point for single Sophie-era block type

toConsensusPointInMode ∷ ConsensusMode mode → ChainPoint → Point (ConsensusBlockForMode mode)

fromConsensusPointInMode ∷ ConsensusMode mode → Point (ConsensusBlockForMode mode) → ChainPoint

Tip of the chain

data ChainTip

This is like a ChainPoint but is conventionally used for the tip of the chain: that is the most recent block at the end of the chain.

It also carries the BlockNo of the chain tip.

Constructors

ChainTipAtGenesis

ChainTip !SlotNo !(Hash BlockHeader) !BlockNo

newtype BlockNo

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

chainTipToChainPoint ∷ ChainTip → ChainPoint

fromConsensusTip ∷ ConsensusBlockForMode mode ~ block ⇒ ConsensusMode mode → Tip block → ChainTip

Data family instances

data family Hash keyrole ∷ Type