Bcc.Api.Certificate

Description

Certificates embedded in transactions

Documentation

data Certificate

Constructors

StakeAddressRegistrationCertificate StakeCredential

StakeAddressDeregistrationCertificate StakeCredential

StakeAddressDelegationCertificate StakeCredential PoolId

StakePoolRegistrationCertificate StakePoolParameters

StakePoolRetirementCertificate PoolId EpochNo

GenesisKeyDelegationCertificate (Hash GenesisKey) (Hash GenesisDelegateKey) (Hash VrfKey)

VestedKeyDelegationCertificate (Hash VestedKey) (Hash VestedDelegateKey) (Hash VrfKey)

MIRCertificate MIRPot MIRTarget

Registering stake address and delegating

makeStakeAddressRegistrationCertificate ∷ StakeCredential → Certificate

makeStakeAddressDeregistrationCertificate ∷ StakeCredential → Certificate

makeStakeAddressDelegationCertificate ∷ StakeCredential → PoolId → Certificate

type PoolId = Hash StakePoolKey

Registering stake pools

makeStakePoolRegistrationCertificate ∷ StakePoolParameters → Certificate

makeStakePoolRetirementCertificate ∷ PoolId → EpochNo → Certificate

data StakePoolParameters

Constructors

StakePoolParameters

Fields

stakePoolId ∷ PoolId

stakePoolVRF ∷ Hash VrfKey

stakePoolCost ∷ Entropic

stakePoolMargin ∷ Rational

stakePoolRewardAccount ∷ StakeAddress

stakePoolPledge ∷ Entropic

stakePoolOwners ∷ [Hash StakeKey]

stakePoolRelays ∷ [StakePoolRelay]

stakePoolMetadata ∷ Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference

Fields

stakePoolMetadataURL ∷ Text

stakePoolMetadataHash ∷ Hash StakePoolMetadata

Special certificates

makeMIRCertificate ∷ MIRPot → MIRTarget → Certificate

makeGenesisKeyDelegationCertificate ∷ Hash GenesisKey → Hash GenesisDelegateKey → Hash VrfKey → Certificate

makeVestedKeyDelegationCertificate ∷ Hash VestedKey → Hash VestedDelegateKey → Hash VrfKey → Certificate

data MIRTarget

The MIRTarget determines the target of a MIRCertificate. A MIRCertificate moves entropic from either the reserves or the treasury to either a collection of stake credentials or to the other pot.

Constructors

StakeAddressesMIR [(StakeCredential, Entropic)]
Use StakeAddressesMIR to make the target of a MIRCertificate a mapping of stake credentials to entropic.

SendToReservesMIR Entropic
Use SendToReservesMIR to make the target of a MIRCertificate the reserves pot.

SendToTreasuryMIR Entropic
Use SendToTreasuryMIR to make the target of a MIRCertificate the treasury pot.

Internal conversion functions

toSophieCertificate ∷ Certificate → DCert StandardCrypto

fromSophieCertificate ∷ DCert StandardCrypto → Certificate

toSophiePoolParams ∷ StakePoolParameters → PoolParams StandardCrypto

fromSophiePoolParams ∷ PoolParams StandardCrypto → StakePoolParameters

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.