Bcc.Api.Cole

Description

This module provides a library interface that is intended to be the complete API for Cole covering everything, including exposing constructors for the lower level types.

Documentation

module Bcc.Api

Cryptographic key interface

data SomeColeSigningKey

Constructors

AColeSigningKeyLegacy (SigningKey ColeKeyLegacy)

AColeSigningKey (SigningKey ColeKey)

Hashes

data family Hash keyrole ∷ Type

Payment addresses

Constructing and inspecting Cole payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Cole addresses, which use the type tag ColeAddr; and

Sophie addresses, which use the type tag SophieAddr. Notably, Sophie addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Cole addresses are the only choice in the Cole era, the Sophie era and all subsequent eras support both Cole and Sophie addresses. The Address type param only says the type of the address (either Cole or Sophie). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

ColeAddress ∷ Address → Address ColeAddr
Cole addresses were the only supported address type in the original Cole era.

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

ColeTxBody ∷ Annotated Tx ByteString → TxBody ColeEra

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

data TxIn

Constructors

TxIn TxId TxIx

data TxOut era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatumHash era)

newtype TxIx

Constructors

TxIx Word

newtype Entropic

Constructors

Entropic Integer

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ColeTx ∷ ATxAux ByteString → Tx ColeEra

Incremental signing and separate witnesses

data KeyWitness era where

Constructors

ColeKeyWitness ∷ TxInWitness → KeyWitness ColeEra

data WitnessNetworkIdOrColeAddress

Either a network ID or a Cole address to be used in constructing a Sophie bootstrap witness.

Constructors

WitnessNetworkId !NetworkId
Network ID.

If this value is used in the construction of a Sophie bootstrap witness, the result will not consist of a derivation path. If that is required, specify a WitnessColeAddress value instead.

WitnessColeAddress !(Address ColeAddr)
Cole address.

If this value is used in the construction of a Sophie bootstrap witness, both the network ID and derivation path will be extracted from the address and used in the construction of the witness.

Errors

class Show e ⇒ Error e where

Methods

displayError ∷ e → String

data FileError e

Constructors

FileError FilePath e

FileErrorTempFile

Fields

FilePath

Target path

FilePath

Temporary path

Handle

FileIOError FilePath IOException

Low level protocol interaction with a Bcc node

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo (ConsensusModeParams mode) NetworkId FilePath

data ColeMode

The Cole-only consensus mode consists of only the Cole era.

This was used on the mainnet before the deployment of the multi-era BccMode. It is now of little practical use, though it illustrates how a single-era consensus mode works. It may be sensible to remove this at some stage.

data ConsensusMode mode where

This GADT provides a value-level representation of all the consensus modes. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

Constructors

ColeMode ∷ ConsensusMode ColeMode

data LocalNodeClientProtocols block point tip tx txerr query m

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols (LocalChainSyncClient block point tip m) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ()))

Chain sync protocol

newtype ChainSyncClient header point tip (m ∷ Type → Type) a

Constructors

ChainSyncClient

Fields

runChainSyncClient ∷ m (ClientStIdle header point tip m a)

Local tx submission

newtype LocalTxSubmissionClient tx reject (m ∷ Type → Type) a

Constructors

LocalTxSubmissionClient (m (LocalTxClientStIdle tx reject m a))

Local state query

newtype LocalStateQueryClient block point (query ∷ Type → Type) (m ∷ Type → Type) a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient ∷ m (ClientStIdle block point query m a)

Address

newtype NetworkMagic

Constructors

NetworkMagic

Fields

unNetworkMagic ∷ Word32

Update Proposal

newtype ColeUpdateProposal

Cole era update proposal

Constructors

ColeUpdateProposal

Fields

unColeUpdateProposal ∷ AProposal ByteString

data ColeProtocolParametersUpdate

Constructors

ColeProtocolParametersUpdate

Fields

bPpuScriptVersion ∷ !(Maybe Word16)

Redundant. This was meant to be the version of the Zerepoch smart contract language, however, there are no smart contracts nor scripts in the Cole era.

bPpuSlotDuration ∷ !(Maybe Natural)

Slot duration in milliseconds.

bPpuMaxBlockSize ∷ !(Maybe Natural)

Maximum block size in bytes.

bPpuMaxHeaderSize ∷ !(Maybe Natural)

Maximum block header size in bytes.

bPpuMaxTxSize ∷ !(Maybe Natural)

Maxiumum transaction size in bytes.

bPpuMaxProposalSize ∷ !(Maybe Natural)

Maximum update proposal size in bytes.

bPpuMpcThd ∷ !(Maybe EntropicPortion)

bPpuHeavyDelThd ∷ !(Maybe EntropicPortion)

Heavyweight delegation threshold. The delegate (i.e stakeholder) must possess no less than this threshold of stake in order to participate in heavyweight delegation.

bPpuUpdateVoteThd ∷ !(Maybe EntropicPortion)

bPpuUpdateProposalThd ∷ !(Maybe EntropicPortion)

bPpuUpdateProposalTTL ∷ !(Maybe SlotNumber)

bPpuSoftforkRule ∷ !(Maybe SoftforkRule)

Values defining the softfork resolution rule. When the stake belonging to block issuers, issuing a given block version, is greater than the current softfork resolution threshold, this block version is adopted.

bPpuTxFeePolicy ∷ !(Maybe TxFeePolicy)

Transaction fee policy represents a formula to compute the minimal allowed Fee for a transaction. Transactions with lesser fees won't be accepted.

bPpuUnlockStakeEpoch ∷ !(Maybe EpochNumber)

This has been re-purposed for unlocking the ShardagnosticBFT logic in the software. Relevant: CDEC-610

makeColeUpdateProposal ∷ NetworkId → ProtocolVersion → SoftwareVersion → SystemTag → InstallerHash → SomeColeSigningKey → ColeProtocolParametersUpdate → ColeUpdateProposal

toColeLedgerUpdateProposal ∷ ColeUpdateProposal → GenTx ColeBlock

makeProtocolParametersUpdate ∷ ColeProtocolParametersUpdate → ProtocolParametersUpdate

Vote

newtype ColeVote

Cole era votes

Constructors

ColeVote

Fields

unColeVote ∷ AVote ByteString

makeColeVote ∷ NetworkId → SomeColeSigningKey → ColeUpdateProposal → Bool → ColeVote

toColeLedgertoColeVote ∷ ColeVote → GenTx ColeBlock

Conversions

fromColeTxIn ∷ TxIn → TxIn

toColeEntropic ∷ Entropic → Maybe Entropic

toColeNetworkMagic ∷ NetworkId → NetworkMagic

toColeProtocolMagicId ∷ NetworkId → ProtocolMagicId

toColeRequiresNetworkMagic ∷ NetworkId → RequiresNetworkMagic