Bcc.Api.Eras

Description

Bcc eras, sometimes we have to distinguish them.

Eras

data ColeEra

A type used as a tag to distinguish the Cole era.

data SophieEra

A type used as a tag to distinguish the Sophie era.

data EvieEra

A type used as a tag to distinguish the Evie era.

data JenEra

A type used as a tag to distinguish the Jen era.

data AurumEra

A type used as a tag to distinguish the Aurum era.

data BccEra era where

This GADT provides a value-level representation of all the Bcc eras. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

This can be used in combination with the IsBccEra class to get access to this value.

In combination this can often enable code that handles all eras, and does so uniformly where possible, and non-uniformly where necessary.

Constructors

ColeEra ∷ BccEra ColeEra

SophieEra ∷ BccEra SophieEra

EvieEra ∷ BccEra EvieEra

JenEra ∷ BccEra JenEra

AurumEra ∷ BccEra AurumEra

class HasTypeProxy era ⇒ IsBccEra era where

The class of Bcc eras. This allows uniform handling of all Bcc eras, but also non-uniform by making case distinctions on the BccEra constructors, or the BccEraStyle constructors via bccEraStyle.

Methods

bccEra ∷ BccEra era

data AnyBccEra where

Constructors

AnyBccEra ∷ IsBccEra era ⇒ BccEra era → AnyBccEra

anyBccEra ∷ BccEra era → AnyBccEra

Like the AnyBccEra constructor but does not demand a IsBccEra class constraint.

data InAnyBccEra thing where

This pairs up some era-dependent type with a BccEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnyBccEra ∷ IsBccEra era ⇒ BccEra era → thing era → InAnyBccEra thing

Deprecated aliases

type Cole = ColeEra

Deprecated: Use ColeEra or ColeAddr as appropriate

type Sophie = SophieEra

Deprecated: Use SophieEra or SophieAddr as appropriate

type Evie = EvieEra

Deprecated: Use EvieEra instead

type Jen = JenEra

Deprecated: Use JenEra instead

Sophie-based eras

data SophieBasedEra era where

While the Cole and Sophie eras are quite different, there are several eras that are based on Sophie with only minor differences. It is useful to be able to treat the Sophie-based eras in a mostly-uniform way.

Values of this type witness the fact that the era is Sophie-based. This can be used to constrain the era to being a Sophie-based on. It allows non-uniform handling making case distinctions on the constructor.

Constructors

SophieBasedEraSophie ∷ SophieBasedEra SophieEra

SophieBasedEraEvie ∷ SophieBasedEra EvieEra

SophieBasedEraJen ∷ SophieBasedEra JenEra

SophieBasedEraAurum ∷ SophieBasedEra AurumEra

class IsBccEra era ⇒ IsSophieBasedEra era where

The class of eras that are based on Sophie. This allows uniform handling of Sophie-based eras, but also non-uniform by making case distinctions on the SophieBasedEra constructors.

Methods

sophieBasedEra ∷ SophieBasedEra era

data InAnySophieBasedEra thing where

This pairs up some era-dependent type with a SophieBasedEra value that tells us what era it is, but hides the era type. This is useful when the era is not statically known, for example when deserialising from a file.

Constructors

InAnySophieBasedEra ∷ IsSophieBasedEra era ⇒ SophieBasedEra era → thing era → InAnySophieBasedEra thing

sophieBasedToBccEra ∷ SophieBasedEra era → BccEra era

Converts a SophieBasedEra to the broader BccEra.

Mapping to era types from the Sophie ledger library

type family SophieLedgerEra era where ...

A type family that connects our era type tags to equivalent type tags used in the Sophie ledger library.

This type mapping connect types from this API with types in the Sophie ledger library which allows writing conversion functions in a more generic way.

Equations

SophieLedgerEra SophieEra = StandardSophie

SophieLedgerEra EvieEra = StandardEvie

SophieLedgerEra JenEra = StandardJen

SophieLedgerEra AurumEra = StandardAurum

Bcc eras, as Cole vs Sophie-based

data BccEraStyle era where

This is the same essential information as BccEra but instead of a flat set of alternative eras, it is factored into the legcy Cole era and the current Sophie-based eras.

This way of factoring the eras is useful because in many cases the major differences are between the Cole and Sophie-based eras, and the Sophie-based eras can often be treated uniformly.

Constructors

LegacyColeEra ∷ BccEraStyle ColeEra

SophieBasedEra ∷ IsSophieBasedEra era ⇒ SophieBasedEra era → BccEraStyle era

bccEraStyle ∷ BccEra era → BccEraStyle era

The BccEraStyle for a BccEra.

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.