Bcc.Api.OperationalCertificate

Description

Operational certificates

Documentation

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter !Word64 !(VerificationKey StakePoolKey)

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod ∷ Word

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)
The stake pool verification key expected for the OperationalCertificateIssueCounter does not match the signing key supplied for signing.

Order: pool vkey expected, pool skey supplied

issueOperationalCertificate ∷ VerificationKey KesKey → Either (SigningKey StakePoolKey) (SigningKey GenesisDelegateExtendedKey) → KESPeriod → OperationalCertificateIssueCounter → Either OperationalCertIssueError (OperationalCertificate, OperationalCertificateIssueCounter)

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.