Bcc.Api.Query

Description

Queries from local clients to the node.

Queries

data QueryInMode mode result where

Constructors

QueryCurrentEra ∷ ConsensusModeIsMultiEra mode → QueryInMode mode AnyBccEra

QueryInEra ∷ EraInMode era mode → QueryInEra era result → QueryInMode mode (Either EraMismatch result)

QueryEraHistory ∷ ConsensusModeIsMultiEra mode → QueryInMode mode (EraHistory mode)

QuerySystemStart ∷ QueryInMode mode SystemStart

data QueryInEra era result where

Constructors

QueryColeUpdateState ∷ QueryInEra ColeEra ColeUpdateState

QueryInSophieBasedEra ∷ SophieBasedEra era → QueryInSophieBasedEra era result → QueryInEra era result

data QueryInSophieBasedEra era result where

Constructors

QueryChainPoint ∷ QueryInSophieBasedEra era ChainPoint

QueryEpoch ∷ QueryInSophieBasedEra era EpochNo

QueryGenesisParameters ∷ QueryInSophieBasedEra era GenesisParameters

QueryProtocolParameters ∷ QueryInSophieBasedEra era ProtocolParameters

QueryProtocolParametersUpdate ∷ QueryInSophieBasedEra era (Map (Hash GenesisKey) ProtocolParametersUpdate)

QueryStakeDistribution ∷ QueryInSophieBasedEra era (Map (Hash StakePoolKey) Rational)

QueryUTxO ∷ QueryUTxOFilter → QueryInSophieBasedEra era (UTxO era)

QueryStakeAddresses ∷ Set StakeCredential → NetworkId → QueryInSophieBasedEra era (Map StakeAddress Entropic, Map StakeAddress PoolId)

QueryStakePools ∷ QueryInSophieBasedEra era (Set PoolId)

QueryStakePoolParameters ∷ Set PoolId → QueryInSophieBasedEra era (Map PoolId StakePoolParameters)

QueryDebugLedgerState ∷ QueryInSophieBasedEra era (SerialisedDebugLedgerState era)

QueryProtocolState ∷ QueryInSophieBasedEra era (ProtocolState era)

data QueryUTxOFilter

Getting the whole UTxO is obviously not efficient since the result can be huge. Filtering by address is also not efficient because it requires a linear search.

The QueryUTxOFilterByTxIn is efficient since it fits with the structure of the UTxO (which is indexed by TxIn).

Constructors

QueryUTxOWhole
O(n) time and space for utxo size n

QueryUTxOByAddress (Set AddressAny)
O(n) time, O(m) space for utxo size n, and address set size m

QueryUTxOByTxIn (Set TxIn)
O(m log n) time, O(m) space for utxo size n, and address set size m

newtype UTxO era

Constructors

UTxO (Map TxIn (TxOut era))

data UTxOInAnyEra where

Constructors

UTxOInAnyEra ∷ BccEra era → UTxO era → UTxOInAnyEra

Internal conversion functions

toConsensusQuery ∷ ∀ mode block result. ConsensusBlockForMode mode ~ block ⇒ QueryInMode mode result → Some (Query block)

fromConsensusQueryResult ∷ ∀ mode block result result'. ConsensusBlockForMode mode ~ block ⇒ QueryInMode mode result → Query block result' → result' → result

Wrapper types used in queries

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (SophieLedgerEra era)))

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (Crypto (SophieLedgerEra era))))

data DebugLedgerState era where

Constructors

DebugLedgerState ∷ SophieLedgerEra era ~ ledgerera ⇒ NewEpochState ledgerera → DebugLedgerState era

data EraHistory mode where

Constructors

EraHistory ∷ ConsensusBlockForMode mode ~ HardForkBlock xs ⇒ ConsensusMode mode → Interpreter xs → EraHistory mode

newtype SystemStart

Constructors

SystemStart

Fields

getSystemStart ∷ UTCTime

newtype SlotsInEpoch

Constructors

SlotsInEpoch Word64

newtype SlotsToEpochEnd

Constructors

SlotsToEpochEnd Word64

slotToEpoch ∷ SlotNo → EraHistory mode → Either PastHorizonException (EpochNo, SlotsInEpoch, SlotsToEpochEnd)

data family LedgerState blk

getProgress ∷ SlotNo → EraHistory mode → Either PastHorizonException (RelativeTime, SlotLength)

Internal conversion functions

toLedgerUTxO ∷ SophieLedgerEra era ~ ledgerera ⇒ Crypto ledgerera ~ StandardCrypto ⇒ SophieBasedEra era → UTxO era → UTxO ledgerera

fromLedgerUTxO ∷ SophieLedgerEra era ~ ledgerera ⇒ Crypto ledgerera ~ StandardCrypto ⇒ SophieBasedEra era → UTxO ledgerera → UTxO era