Bcc.Api.Script

Languages

data SimpleScriptV1

The original simple script language which supports

require a signature from a given key (by verification key hash)

n-way and combinator

n-way or combinator

m-of-n combinator

This version of the language was introduced in the SophieEra.

data SimpleScriptV2

The second version of the simple script language. It has all the features of SimpleScriptV1 plus new atomic predicates:

require the time be before a given slot number

require the time be after a given slot number

This version of the language was introduced in the EvieEra.

data ZerepochScriptV1

Place holder type to show what the pattern is to extend to multiple languages, not just multiple versions of a single language.

data ScriptLanguage lang where

Constructors

SimpleScriptLanguage ∷ SimpleScriptVersion lang → ScriptLanguage lang

ZerepochScriptLanguage ∷ ZerepochScriptVersion lang → ScriptLanguage lang

data SimpleScriptVersion lang where

Constructors

SimpleScriptV1 ∷ SimpleScriptVersion SimpleScriptV1

SimpleScriptV2 ∷ SimpleScriptVersion SimpleScriptV2

data ZerepochScriptVersion lang where

Constructors

ZerepochScriptV1 ∷ ZerepochScriptVersion ZerepochScriptV1

data AnyScriptLanguage where

Constructors

AnyScriptLanguage ∷ ScriptLanguage lang → AnyScriptLanguage

data AnyZerepochScriptVersion where

Constructors

AnyZerepochScriptVersion ∷ ZerepochScriptVersion lang → AnyZerepochScriptVersion

class HasTypeProxy lang ⇒ IsScriptLanguage lang where

Methods

scriptLanguage ∷ ScriptLanguage lang

class IsScriptLanguage lang ⇒ IsSimpleScriptLanguage lang where

Methods

simpleScriptVersion ∷ SimpleScriptVersion lang

Scripts in a specific language

data Script lang where

A script in a particular language.

See also ScriptInAnyLang for a script in any of the known languages.

See also ScriptInEra for a script in a language that is available within a particular era.

Note that some but not all scripts have an external JSON syntax, hence this type has no JSON serialisation instances. The SimpleScript family of languages do have a JSON syntax and thus have ToJSON/FromJSON instances.

Constructors

SimpleScript ∷ !(SimpleScriptVersion lang) → !(SimpleScript lang) → Script lang

ZerepochScript ∷ !(ZerepochScriptVersion lang) → !(ZerepochScript lang) → Script lang

Scripts in any language

data ScriptInAnyLang where

Sometimes it is necessary to handle all languages without making static type distinctions between languages. For example, when reading external input, or before the era context is known.

Use toScriptInEra to convert to a script in the context of an era.

Constructors

ScriptInAnyLang ∷ ScriptLanguage lang → Script lang → ScriptInAnyLang

toScriptInAnyLang ∷ Script lang → ScriptInAnyLang

Convert a script in a specific statically-known language to a ScriptInAnyLang.

No inverse to this is provided, just do case analysis on the ScriptLanguage field within the ScriptInAnyLang constructor.

Scripts in an era

data ScriptInEra era where

Constructors

ScriptInEra ∷ ScriptLanguageInEra lang era → Script lang → ScriptInEra era

toScriptInEra ∷ BccEra era → ScriptInAnyLang → Maybe (ScriptInEra era)

Given a target era and a script in some language, check if the language is supported in that era, and if so return a ScriptInEra.

eraOfScriptInEra ∷ ScriptInEra era → SophieBasedEra era

Use of a script in an era as a witness

data WitCtxTxIn

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness a transaction input.

data WitCtxMint

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness minting.

data WitCtxStake

A tag type for the context in which a script is used in a transaction.

This type tags the context as being to witness the use of stake addresses in both certificates and withdrawals.

data WitCtx witctx where

This GADT provides a value-level representation of all the witness contexts. This enables pattern matching on the context to allow them to be treated in a non-uniform way.

Constructors

WitCtxTxIn ∷ WitCtx WitCtxTxIn

WitCtxMint ∷ WitCtx WitCtxMint

WitCtxStake ∷ WitCtx WitCtxStake

data ScriptWitness witctx era where

A use of a script within a transaction body to witness that something is being used in an authorised manner. That can be

spending a transaction input

minting tokens

using a certificate (stake address certs specifically)

withdrawing from a reward account

For simple script languages, the use of the script is the same in all contexts. For Zerepoch scripts, using a script involves supplying a redeemer. In addition, Zerepoch scripts used for spending inputs must also supply the datum value used when originally creating the TxOut that is now being spent.

Constructors

SimpleScriptWitness ∷ ScriptLanguageInEra lang era → SimpleScriptVersion lang → SimpleScript lang → ScriptWitness witctx era

ZerepochScriptWitness ∷ ScriptLanguageInEra lang era → ZerepochScriptVersion lang → ZerepochScript lang → ScriptDatum witctx → ScriptRedeemer → ExecutionUnits → ScriptWitness witctx era

data Witness witctx era where

Constructors

KeyWitness ∷ KeyWitnessInCtx witctx → Witness witctx era

ScriptWitness ∷ ScriptWitnessInCtx witctx → ScriptWitness witctx era → Witness witctx era

data KeyWitnessInCtx witctx where

Constructors

KeyWitnessForSpending ∷ KeyWitnessInCtx WitCtxTxIn

KeyWitnessForStakeAddr ∷ KeyWitnessInCtx WitCtxStake

data ScriptWitnessInCtx witctx where

Constructors

ScriptWitnessForSpending ∷ ScriptWitnessInCtx WitCtxTxIn

ScriptWitnessForMinting ∷ ScriptWitnessInCtx WitCtxMint

ScriptWitnessForStakeAddr ∷ ScriptWitnessInCtx WitCtxStake

data ScriptDatum witctx where

Constructors

ScriptDatumForTxIn ∷ ScriptData → ScriptDatum WitCtxTxIn

NoScriptDatumForMint ∷ ScriptDatum WitCtxMint

NoScriptDatumForStake ∷ ScriptDatum WitCtxStake

type ScriptRedeemer = ScriptData

scriptWitnessScript ∷ ScriptWitness witctx era → ScriptInEra era

Languages supported in each era

data ScriptLanguageInEra lang era where

Constructors

SimpleScriptV1InSophie ∷ ScriptLanguageInEra SimpleScriptV1 SophieEra

SimpleScriptV1InEvie ∷ ScriptLanguageInEra SimpleScriptV1 EvieEra

SimpleScriptV1InJen ∷ ScriptLanguageInEra SimpleScriptV1 JenEra

SimpleScriptV1InAurum ∷ ScriptLanguageInEra SimpleScriptV1 AurumEra

SimpleScriptV2InEvie ∷ ScriptLanguageInEra SimpleScriptV2 EvieEra

SimpleScriptV2InJen ∷ ScriptLanguageInEra SimpleScriptV2 JenEra

SimpleScriptV2InAurum ∷ ScriptLanguageInEra SimpleScriptV2 AurumEra

ZerepochScriptV1InAurum ∷ ScriptLanguageInEra ZerepochScriptV1 AurumEra

scriptLanguageSupportedInEra ∷ BccEra era → ScriptLanguage lang → Maybe (ScriptLanguageInEra lang era)

Check if a given script language is supported in a given era, and if so return the evidence.

languageOfScriptLanguageInEra ∷ ScriptLanguageInEra lang era → ScriptLanguage lang

eraOfScriptLanguageInEra ∷ ScriptLanguageInEra lang era → SophieBasedEra era

The simple script language

data SimpleScript lang where

Constructors

RequireSignature ∷ !(Hash PaymentKey) → SimpleScript lang

RequireTimeBefore ∷ !(TimeLocksSupported lang) → !SlotNo → SimpleScript lang

RequireTimeAfter ∷ !(TimeLocksSupported lang) → !SlotNo → SimpleScript lang

RequireAllOf ∷ [SimpleScript lang] → SimpleScript lang

RequireAnyOf ∷ [SimpleScript lang] → SimpleScript lang

RequireMOf ∷ Int → [SimpleScript lang] → SimpleScript lang

data TimeLocksSupported lang where

Time lock feature in the SimpleScript language.

The constructors of this type serve as evidence that the timelocks feature is supported in particular versions of the language.

Constructors

TimeLocksInSimpleScriptV2 ∷ TimeLocksSupported SimpleScriptV2

timeLocksSupported ∷ SimpleScriptVersion lang → Maybe (TimeLocksSupported lang)

adjustSimpleScriptVersion ∷ SimpleScriptVersion lang' → SimpleScript lang → Maybe (SimpleScript lang')

Try converting the SimpleScript into a different version of the language.

This will work when the script only uses the features of the target language version. For example converting from SimpleScriptV2 to SimpleScriptV1 will work if the script happens not to use time locks feature. On the other hand converting SimpleScriptV1 to SimpleScriptV2 will always work because it is backwards compatible.

The Zerepoch script language

data ZerepochScript lang where

Zerepoch scripts.

Note that Zerepoch scripts have a binary serialisation but no JSON serialisation.

Constructors

ZerepochScriptSerialised ∷ ShortByteString → ZerepochScript lang

exampleZerepochScriptAlwaysSucceeds ∷ WitCtx witctx → ZerepochScript ZerepochScriptV1

An example Zerepoch script that always succeeds, irrespective of inputs.

For example, if one were to use this for a payment address then it would allow anyone to spend from it.

The exact script depends on the context in which it is to be used.

exampleZerepochScriptAlwaysFails ∷ WitCtx witctx → ZerepochScript ZerepochScriptV1

An example Zerepoch script that always fails, irrespective of inputs.

For example, if one were to use this for a payment address then it would be impossible for anyone to ever spend from it.

The exact script depends on the context in which it is to be used.

Script data

data ScriptData

Constructors

ScriptDataConstructor Integer [ScriptData]

ScriptDataMap [(ScriptData, ScriptData)]

ScriptDataList [ScriptData]

ScriptDataNumber Integer

ScriptDataBytes ByteString

Script execution units

data ExecutionUnits

The units for how long a script executes for and how much memory it uses. This is used to declare the resources used by a particular use of a script.

This type is also used to describe the limits for the maximum overall execution units per transaction or per block.

Constructors

ExecutionUnits

Fields

executionSteps ∷ Word64

This corresponds roughly to the time to execute a script.

executionMemory ∷ Word64

This corresponds roughly to the peak memory used during script execution.

Script hashes

newtype ScriptHash

We have this type separate from the Hash type to avoid the script hash type being parametrised by the era. The representation is era independent, and there are many places where we want to use a script hash where we don't want things to be era-parametrised.

Constructors

ScriptHash (ScriptHash StandardCrypto)

hashScript ∷ Script lang → ScriptHash

Internal conversion functions

toSophieScript ∷ ScriptInEra era → Script (SophieLedgerEra era)

fromSophieBasedScript ∷ SophieBasedEra era → Script (SophieLedgerEra era) → ScriptInEra era

toSophieMultiSig ∷ SimpleScript SimpleScriptV1 → MultiSig StandardCrypto

Conversion for the MultiSig language used by the Sophie era.

fromSophieMultiSig ∷ MultiSig StandardCrypto → SimpleScript lang

Conversion for the MultiSig language used by the Sophie era.

toEvieTimelock ∷ ∀ lang. SimpleScript lang → Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Evie and Jen eras.

fromEvieTimelock ∷ TimeLocksSupported lang → Timelock StandardCrypto → SimpleScript lang

Conversion for the Timelock language that is shared between the Evie and Jen eras.

toAurumExUnits ∷ ExecutionUnits → ExUnits

fromAurumExUnits ∷ ExUnits → ExecutionUnits

toSophieScriptHash ∷ ScriptHash → ScriptHash StandardCrypto

fromSophieScriptHash ∷ ScriptHash StandardCrypto → ScriptHash

toZerepochData ∷ ScriptData → Data

fromZerepochData ∷ Data → ScriptData

toAurumData ∷ ScriptData → Data ledgerera

fromAurumData ∷ Data ledgerera → ScriptData

toAurumLanguage ∷ AnyZerepochScriptVersion → Language

fromAurumLanguage ∷ Language → AnyZerepochScriptVersion

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

data family Hash keyrole ∷ Type