Bcc.Api.Sophie

Description

This module provides a library interface that is intended to be the complete API for Sophie covering everything, including exposing constructors for the lower level types.

Documentation

module Bcc.Api

Genesis

data SophieGenesis era

Constructors

SophieGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !Word64

sgVestMultiple ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxEntropicSupply ∷ !Word64

sgProtocolParams ∷ !(PParams era)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgVestedDelegs ∷ !(Map (KeyHash 'Vested (Crypto era)) (VestedDelegPair (Crypto era)))

sgInitialFunds ∷ !(Map (Addr (Crypto era)) Coin)

sgStaking ∷ !(SophieGenesisStaking (Crypto era))

sophieGenesisDefaults ∷ SophieGenesis crypto

Some reasonable starting defaults for constructing a SophieGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxEntropicSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.

Cryptographic key interface

Hashes

data family Hash keyrole ∷ Type

Payment addresses

Constructing and inspecting Sophie payment addresses

data Address addrtype where

Addresses are used as locations where assets live. The address determines the rights needed to spend assets at the address: in particular holding some signing key or being able to satisfy the conditions of a script.

There are currently two types of address:

Cole addresses, which use the type tag ColeAddr; and

Sophie addresses, which use the type tag SophieAddr. Notably, Sophie addresses support scripts and stake delegation.

The address type is subtly from the ledger era in which each address type is valid: while Cole addresses are the only choice in the Cole era, the Sophie era and all subsequent eras support both Cole and Sophie addresses. The Address type param only says the type of the address (either Cole or Sophie). The AddressInEra type connects the address type with the era in which it is supported.

Constructors

SophieAddress ∷ Network → PaymentCredential StandardCrypto → StakeReference StandardCrypto → Address SophieAddr
Sophie addresses allow delegation. Sophie addresses were introduced in Sophie era and are thus supported from the Sophie era onwards

toSophieAddr ∷ AddressInEra era → Addr StandardCrypto

fromSophieAddr ∷ IsSophieBasedEra era ⇒ Addr StandardCrypto → AddressInEra era

toSophieStakeCredential ∷ StakeCredential → StakeCredential StandardCrypto

fromSophieStakeCredential ∷ StakeCredential StandardCrypto → StakeCredential

data NetworkId

Constructors

Mainnet

Testnet !NetworkMagic

Stake addresses

data PaymentCredential

Constructors

PaymentCredentialByKey (Hash PaymentKey)

PaymentCredentialByScript ScriptHash

data StakeAddress where

Constructors

StakeAddress ∷ Network → StakeCredential StandardCrypto → StakeAddress

data StakeAddressReference

Constructors

StakeAddressByValue StakeCredential

StakeAddressByPointer StakeAddressPointer

NoStakeAddress

data StakeCredential

Constructors

StakeCredentialByKey (Hash StakeKey)

StakeCredentialByScript ScriptHash

toSophieStakeAddr ∷ StakeAddress → RewardAcnt StandardCrypto

fromSophieStakeAddr ∷ RewardAcnt StandardCrypto → StakeAddress

fromSophieStakeReference ∷ StakeReference StandardCrypto → StakeAddressReference

fromSophiePaymentCredential ∷ PaymentCredential StandardCrypto → PaymentCredential

Building transactions

Constructing and inspecting transactions

data TxBody era where

Constructors

SophieTxBody

Fields

∷ SophieBasedEra era

→ TxBody (SophieLedgerEra era)

→ [Script (SophieLedgerEra era)]

→ TxBodyScriptData era

→ Maybe (AuxiliaryData (SophieLedgerEra era))

→ TxScriptValidity era

Mark script as expected to pass or fail validation

→ TxBody era

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

toSophieTxId ∷ TxId → TxId StandardCrypto

fromSophieTxId ∷ TxId StandardCrypto → TxId

getTxIdSophie ∷ Crypto (SophieLedgerEra era) ~ StandardCrypto ⇒ UsesTxBody (SophieLedgerEra era) ⇒ SophieBasedEra era → TxBody (SophieLedgerEra era) → TxId

data TxIn

Constructors

TxIn TxId TxIx

toSophieTxIn ∷ TxIn → TxIn StandardCrypto

fromSophieTxIn ∷ TxIn StandardCrypto → TxIn

data TxOut era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatumHash era)

toSophieTxOut ∷ ∀ era ledgerera. SophieLedgerEra era ~ ledgerera ⇒ SophieBasedEra era → TxOut era → TxOut ledgerera

fromSophieTxOut ∷ SophieLedgerEra era ~ ledgerera ⇒ SophieBasedEra era → TxOut ledgerera → TxOut era

newtype TxIx

Constructors

TxIx Word

newtype Entropic

Constructors

Entropic Integer

toSophieEntropic ∷ Entropic → Coin

fromSophieEntropic ∷ Coin → Entropic

toJenValue ∷ Value → Value StandardCrypto

fromJenValue ∷ Value StandardCrypto → Value

calcMinimumDeposit ∷ Value → Entropic → Entropic

Calculate cost of making a UTxO entry for a given Value and mininimum UTxO value derived from the ProtocolParameters

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

SophieTx ∷ SophieBasedEra era → Tx (SophieLedgerEra era) → Tx era

Incremental signing and separate witnesses

data KeyWitness era where

Constructors

SophieBootstrapWitness ∷ SophieBasedEra era → BootstrapWitness StandardCrypto → KeyWitness era

SophieKeyWitness ∷ SophieBasedEra era → WitVKey Witness StandardCrypto → KeyWitness era

data SophieWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

WitnessVestedKey (SigningKey VestedKey)

WitnessVestedExtendedKey (SigningKey VestedExtendedKey)

WitnessVestedDelegateKey (SigningKey VestedDelegateKey)

WitnessVestedDelegateExtendedKey (SigningKey VestedDelegateExtendedKey)

data SophieSigningKey

We support making key witnesses with both normal and extended signing keys.

Constructors

SophieNormalSigningKey (SignKeyDSIGN StandardCrypto)
A normal ed25519 signing key

SophieExtendedSigningKey XPrv
An extended ed25519 signing key

getSophieKeyWitnessVerificationKey ∷ SophieSigningKey → VKey Witness StandardCrypto

makeSophieSignature ∷ SignableRepresentation tosign ⇒ tosign → SophieSigningKey → SignedDSIGN StandardCrypto tosign

toSophieSigningKey ∷ SophieWitnessSigningKey → SophieSigningKey

Transaction metadata

Embedding additional structured data within transactions.

toSophieMetadata ∷ Map Word64 TxMetadataValue → Map Word64 Metadatum

fromSophieMetadata ∷ Map Word64 Metadatum → Map Word64 TxMetadataValue

toSophieMetadatum ∷ TxMetadataValue → Metadatum

fromSophieMetadatum ∷ Metadatum → TxMetadataValue

Protocol parameters

data ProtocolParameters

The values of the set of updateable protocol paramaters. At any particular point on the chain there is a current set of paramaters in use.

These paramaters can be updated (at epoch boundaries) via an UpdateProposal, which contains a ProtocolParametersUpdate.

The ProtocolParametersUpdate is essentially a diff for the ProtocolParameters.

There are also paramaters fixed in the Genesis file. See GenesisParameters.

Constructors

ProtocolParameters

Fields

protocolParamProtocolVersion ∷ (Natural, Natural)

Protocol version, major and minor. Updating the major version is used to trigger hard forks.

protocolParamDecentralization ∷ Rational

The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Optimum schedule. So 1 means fully centralised, while 0 means fully decentralised.

This is the "d" parameter from the design document.

protocolParamExtraOptimumEntropy ∷ Maybe OptimumNonce

Extra entropy for the Optimum per-epoch nonce.

This can be used to add extra entropy during the decentralisation process. If the extra entropy can be demonstrated to be generated randomly then this method can be used to show that the initial federated operators did not subtly bias the initial schedule so that they retain undue influence after decentralisation.

protocolParamMaxBlockHeaderSize ∷ Natural

The maximum permitted size of a block header.

This must be at least as big as the largest legitimate block headers but should not be too much larger, to help prevent DoS attacks.

Caution: setting this to be smaller than legitimate block headers is a sure way to brick the system!

protocolParamMaxBlockBodySize ∷ Natural

The maximum permitted size of the block body (that is, the block payload, without the block header).

This should be picked with the Optimum network delta security parameter in mind. Making this too large can severely weaken the Optimum consensus properties.

Caution: setting this to be smaller than a transaction that can change the protocol parameters is a sure way to brick the system!

protocolParamMaxTxSize ∷ Natural

The maximum permitted size of a transaction.

Typically this should not be too high a fraction of the block size, otherwise wastage from block fragmentation becomes a problem, and the current implementation does not use any sophisticated box packing algorithm.

protocolParamTxFeeFixed ∷ Natural

The constant factor for the minimum fee calculation.

protocolParamTxFeePerByte ∷ Natural

The linear factor for the minimum fee calculation.

protocolParamMinUTxOValue ∷ Maybe Entropic

The minimum permitted value for new UTxO entries, ie for transaction outputs.

protocolParamStakeAddressDeposit ∷ Entropic

The deposit required to register a stake address.

protocolParamStakePoolDeposit ∷ Entropic

The deposit required to register a stake pool.

protocolParamMinPoolCost ∷ Entropic

The minimum value that stake pools are permitted to declare for their cost parameter.

protocolParamPoolRetireMaxEpoch ∷ EpochNo

The maximum number of epochs into the future that stake pools are permitted to schedule a retirement.

protocolParamStakePoolTargetNum ∷ Natural

The equilibrium target number of stake pools.

This is the "k" incentives parameter from the design document.

protocolParamPoolPledgeInfluence ∷ Rational

The influence of the pledge in stake pool rewards.

This is the "a_0" incentives parameter from the design document.

protocolParamMonetaryExpansion ∷ Rational

The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch.

This is the "rho" incentives parameter from the design document.

protocolParamTreasuryCut ∷ Rational

The fraction of the fee pot each epoch that goes to the treasury.

This is the "tau" incentives parameter from the design document.

protocolParamUTxOCostPerWord ∷ Maybe Entropic

Cost in bcc per word of UTxO storage.

Introduced in Aurum

protocolParamCostModels ∷ Map AnyZerepochScriptVersion CostModel

Cost models for script languages that use them.

Introduced in Aurum

protocolParamPrices ∷ Maybe ExecutionUnitPrices

Price of execution units for script languages that use them.

Introduced in Aurum

protocolParamMaxTxExUnits ∷ Maybe ExecutionUnits

Max total script execution resources units allowed per tx

Introduced in Aurum

protocolParamMaxBlockExUnits ∷ Maybe ExecutionUnits

Max total script execution resources units allowed per block

Introduced in Aurum

protocolParamMaxValueSize ∷ Maybe Natural

Max size of a Value in a tx ouput.

Introduced in Aurum

protocolParamCollateralPercent ∷ Maybe Natural

The percentage of the script contribution to the txfee that must be provided as collateral inputs when including Zerepoch scripts.

Introduced in Aurum

protocolParamMaxCollateralInputs ∷ Maybe Natural

The maximum number of collateral inputs allowed in a transaction.

Introduced in Aurum

checkProtocolParameters ∷ ∀ era. IsBccEra era ⇒ SophieBasedEra era → ProtocolParameters → Either ProtocolParametersError ()

data ProtocolParametersError

Constructors

PParamsErrorMissingMinUTxoValue AnyBccEra

PParamsErrorMissingAurumProtocolParameter

Scripts

toSophieScript ∷ ScriptInEra era → Script (SophieLedgerEra era)

toSophieMultiSig ∷ SimpleScript SimpleScriptV1 → MultiSig StandardCrypto

Conversion for the MultiSig language used by the Sophie era.

fromSophieMultiSig ∷ MultiSig StandardCrypto → SimpleScript lang

Conversion for the MultiSig language used by the Sophie era.

toEvieTimelock ∷ ∀ lang. SimpleScript lang → Timelock StandardCrypto

Conversion for the Timelock language that is shared between the Evie and Jen eras.

fromEvieTimelock ∷ TimeLocksSupported lang → Timelock StandardCrypto → SimpleScript lang

Conversion for the Timelock language that is shared between the Evie and Jen eras.

toSophieScriptHash ∷ ScriptHash → ScriptHash StandardCrypto

fromSophieScriptHash ∷ ScriptHash StandardCrypto → ScriptHash

data ZerepochScript lang where

Zerepoch scripts.

Note that Zerepoch scripts have a binary serialisation but no JSON serialisation.

Constructors

ZerepochScriptSerialised ∷ ShortByteString → ZerepochScript lang

toZerepochData ∷ ScriptData → Data

fromZerepochData ∷ Data → ScriptData

toAurumData ∷ ScriptData → Data ledgerera

fromAurumData ∷ Data ledgerera → ScriptData

toAurumPrices ∷ ExecutionUnitPrices → Maybe Prices

fromAurumPrices ∷ Prices → ExecutionUnitPrices

toAurumExUnits ∷ ExecutionUnits → ExUnits

fromAurumExUnits ∷ ExUnits → ExecutionUnits

toAurumRdmrPtr ∷ ScriptWitnessIndex → RdmrPtr

fromAurumRdmrPtr ∷ RdmrPtr → ScriptWitnessIndex

Certificates

data Certificate

Constructors

StakeAddressRegistrationCertificate StakeCredential

StakeAddressDeregistrationCertificate StakeCredential

StakeAddressDelegationCertificate StakeCredential PoolId

StakePoolRegistrationCertificate StakePoolParameters

StakePoolRetirementCertificate PoolId EpochNo

GenesisKeyDelegationCertificate (Hash GenesisKey) (Hash GenesisDelegateKey) (Hash VrfKey)

VestedKeyDelegationCertificate (Hash VestedKey) (Hash VestedDelegateKey) (Hash VrfKey)

MIRCertificate MIRPot MIRTarget

toSophieCertificate ∷ Certificate → DCert StandardCrypto

fromSophieCertificate ∷ DCert StandardCrypto → Certificate

Operational certificates

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter !Word64 !(VerificationKey StakePoolKey)

data OperationalCertIssueError

Constructors

OperationalCertKeyMismatch (VerificationKey StakePoolKey) (VerificationKey StakePoolKey)
The stake pool verification key expected for the OperationalCertificateIssueCounter does not match the signing key supplied for signing.

Order: pool vkey expected, pool skey supplied

Stake Pool

data StakePoolMetadata

A representation of the required fields for off-chain stake pool metadata.

Constructors

StakePoolMetadata !Text !Text !Text !Text

stakePoolName ∷ StakePoolMetadata → Text

A name of up to 50 characters.

stakePoolDescription ∷ StakePoolMetadata → Text

A description of up to 255 characters.

stakePoolTicker ∷ StakePoolMetadata → Text

A ticker of 3-5 characters, for a compact display of stake pools in a wallet.

stakePoolHomepage ∷ StakePoolMetadata → Text

A URL to a homepage with additional information about the pool. n.b. the spec does not specify a character limit for this field.

data StakePoolMetadataReference

Constructors

StakePoolMetadataReference Text (Hash StakePoolMetadata)

stakePoolMetadataURL ∷ StakePoolMetadataReference → Text

stakePoolMetadataHash ∷ StakePoolMetadataReference → Hash StakePoolMetadata

data StakePoolParameters

Constructors

StakePoolParameters PoolId (Hash VrfKey) Entropic Rational StakeAddress Entropic [Hash StakeKey] [StakePoolRelay] (Maybe StakePoolMetadataReference)

stakePoolId ∷ StakePoolParameters → PoolId

stakePoolVRF ∷ StakePoolParameters → Hash VrfKey

stakePoolCost ∷ StakePoolParameters → Entropic

stakePoolMargin ∷ StakePoolParameters → Rational

stakePoolRewardAccount ∷ StakePoolParameters → StakeAddress

stakePoolPledge ∷ StakePoolParameters → Entropic

stakePoolOwners ∷ StakePoolParameters → [Hash StakeKey]

stakePoolRelays ∷ StakePoolParameters → [StakePoolRelay]

stakePoolMetadata ∷ StakePoolParameters → Maybe StakePoolMetadataReference

data StakePoolRelay

Constructors

StakePoolRelayIp (Maybe IPv4) (Maybe IPv6) (Maybe PortNumber)
One or both of IPv4 & IPv6

StakePoolRelayDnsARecord ByteString (Maybe PortNumber)
An DNS name pointing to a A or AAAA record.

StakePoolRelayDnsSrvRecord ByteString
A DNS name pointing to a SRV record.

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

Stake pool operator's keys

data StakePoolKey

type PoolId = Hash StakePoolKey

KES keys

data KesKey

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod ∷ Word

VRF keys

data VrfKey

Low level protocol interaction with a Bcc node

data LocalNodeConnectInfo mode

Constructors

LocalNodeConnectInfo (ConsensusModeParams mode) NetworkId FilePath

data SophieMode

The Sophie-only consensus mode consists of only the Sophie era.

This was used for the early Sophie testnets prior to the use of the multi-era BccMode. It is useful for setting up Sophie test networks (e.g. for benchmarking) without having to go through the complication of the hard fork from Cole to Sophie eras. It also shows how a single-era consensus mode works. It may be replaced by other single-era modes in future.

data ConsensusMode mode where

This GADT provides a value-level representation of all the consensus modes. This enables pattern matching on the era to allow them to be treated in a non-uniform way.

Constructors

SophieMode ∷ ConsensusMode SophieMode

data LocalNodeClientProtocols block point tip tx txerr query m

The protocols we can use with a local node. Use in conjunction with connectToLocalNode.

These protocols use the types from the rest of this API. The conversion to/from the types used by the underlying wire formats is handled by connectToLocalNode.

Constructors

LocalNodeClientProtocols (LocalChainSyncClient block point tip m) (Maybe (LocalTxSubmissionClient tx txerr m ())) (Maybe (LocalStateQueryClient block point query m ()))

Sophie based eras

type family SophieLedgerEra era where ...

A type family that connects our era type tags to equivalent type tags used in the Sophie ledger library.

This type mapping connect types from this API with types in the Sophie ledger library which allows writing conversion functions in a more generic way.

Equations

SophieLedgerEra SophieEra = StandardSophie

SophieLedgerEra EvieEra = StandardEvie

SophieLedgerEra JenEra = StandardJen

SophieLedgerEra AurumEra = StandardAurum

Local State Query

data DebugLedgerState era where

Constructors

DebugLedgerState ∷ SophieLedgerEra era ~ ledgerera ⇒ NewEpochState ledgerera → DebugLedgerState era

newtype ProtocolState era

Constructors

ProtocolState (Serialised (ChainDepState (Crypto (SophieLedgerEra era))))

newtype SerialisedDebugLedgerState era

Constructors

SerialisedDebugLedgerState (Serialised (NewEpochState (SophieLedgerEra era)))

newtype UTxO era

Constructors

UTxO (Map TxIn (TxOut era))

Conversions

toSophieNetwork ∷ NetworkId → Network

fromSophiePParams ∷ PParams ledgerera → ProtocolParameters