Bcc.Api.Sophie.Genesis

Documentation

data SophieGenesis era

Constructors

SophieGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !PositiveUnitInterval

sgSecurityParam ∷ !Word64

sgVestMultiple ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxEntropicSupply ∷ !Word64

sgProtocolParams ∷ !(PParams era)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgVestedDelegs ∷ !(Map (KeyHash 'Vested (Crypto era)) (VestedDelegPair (Crypto era)))

sgInitialFunds ∷ !(Map (Addr (Crypto era)) Coin)

sgStaking ∷ !(SophieGenesisStaking (Crypto era))

sophieGenesisDefaults ∷ SophieGenesis crypto

Some reasonable starting defaults for constructing a SophieGenesis.

You must override at least the following fields for this to be useful:

sgSystemStart the time of the first block

sgNetworkMagic to a suitable testnet or mainnet network magic number.

sgGenDelegs to have some initial nodes

sgInitialFunds to have any money in the system

sgMaxEntropicSupply must be at least the sum of the sgInitialFunds but more if you want to allow for rewards.