Bcc.Api.Tx

Description

Complete, signed transactions

Signing transactions

Creating transaction witnesses one by one, or all in one go.

data Tx era where

Constructors

ColeTx ∷ ATxAux ByteString → Tx ColeEra

SophieTx ∷ SophieBasedEra era → Tx (SophieLedgerEra era) → Tx era

Bundled Patterns

pattern Tx ∷ TxBody era → [KeyWitness era] → Tx era

getTxBody ∷ ∀ era. Tx era → TxBody era

getTxWitnesses ∷ ∀ era. Tx era → [KeyWitness era]

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

Signing in one go

data SophieSigningKey

We support making key witnesses with both normal and extended signing keys.

Constructors

SophieNormalSigningKey (SignKeyDSIGN StandardCrypto)
A normal ed25519 signing key

SophieExtendedSigningKey XPrv
An extended ed25519 signing key

toSophieSigningKey ∷ SophieWitnessSigningKey → SophieSigningKey

signColeTransaction ∷ NetworkId → TxBody ColeEra → [SigningKey ColeKey] → Tx ColeEra

signSophieTransaction ∷ IsSophieBasedEra era ⇒ TxBody era → [SophieWitnessSigningKey] → Tx era

Incremental signing and separate witnesses

makeSignedTransaction ∷ ∀ era. [KeyWitness era] → TxBody era → Tx era

data KeyWitness era where

Constructors

ColeKeyWitness ∷ TxInWitness → KeyWitness ColeEra

SophieBootstrapWitness ∷ SophieBasedEra era → BootstrapWitness StandardCrypto → KeyWitness era

SophieKeyWitness ∷ SophieBasedEra era → WitVKey Witness StandardCrypto → KeyWitness era

makeColeKeyWitness ∷ ∀ key. IsColeKey key ⇒ NetworkId → TxBody ColeEra → SigningKey key → KeyWitness ColeEra

data SophieWitnessSigningKey

Constructors

WitnessPaymentKey (SigningKey PaymentKey)

WitnessPaymentExtendedKey (SigningKey PaymentExtendedKey)

WitnessStakeKey (SigningKey StakeKey)

WitnessStakeExtendedKey (SigningKey StakeExtendedKey)

WitnessStakePoolKey (SigningKey StakePoolKey)

WitnessGenesisKey (SigningKey GenesisKey)

WitnessGenesisExtendedKey (SigningKey GenesisExtendedKey)

WitnessGenesisDelegateKey (SigningKey GenesisDelegateKey)

WitnessGenesisDelegateExtendedKey (SigningKey GenesisDelegateExtendedKey)

WitnessGenesisUTxOKey (SigningKey GenesisUTxOKey)

WitnessVestedKey (SigningKey VestedKey)

WitnessVestedExtendedKey (SigningKey VestedExtendedKey)

WitnessVestedDelegateKey (SigningKey VestedDelegateKey)

WitnessVestedDelegateExtendedKey (SigningKey VestedDelegateExtendedKey)

WitnessVestedUTxOKey (SigningKey VestedUTxOKey)

makeSophieKeyWitness ∷ ∀ era. IsSophieBasedEra era ⇒ TxBody era → SophieWitnessSigningKey → KeyWitness era

data WitnessNetworkIdOrColeAddress

Either a network ID or a Cole address to be used in constructing a Sophie bootstrap witness.

Constructors

WitnessNetworkId !NetworkId
Network ID.

If this value is used in the construction of a Sophie bootstrap witness, the result will not consist of a derivation path. If that is required, specify a WitnessColeAddress value instead.

WitnessColeAddress !(Address ColeAddr)
Cole address.

If this value is used in the construction of a Sophie bootstrap witness, both the network ID and derivation path will be extracted from the address and used in the construction of the witness.

makeSophieBootstrapWitness ∷ ∀ era. IsSophieBasedEra era ⇒ WitnessNetworkIdOrColeAddress → TxBody era → SigningKey ColeKey → KeyWitness era

makeSophieSignature ∷ SignableRepresentation tosign ⇒ tosign → SophieSigningKey → SignedDSIGN StandardCrypto tosign

getSophieKeyWitnessVerificationKey ∷ SophieSigningKey → VKey Witness StandardCrypto

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.