Bcc.Api.TxBody

Description

Transaction bodies

Transaction bodies

data TxBody era where

Constructors

ColeTxBody ∷ Annotated Tx ByteString → TxBody ColeEra

SophieTxBody

Fields

∷ SophieBasedEra era

→ TxBody (SophieLedgerEra era)

→ [Script (SophieLedgerEra era)]

→ TxBodyScriptData era

→ Maybe (AuxiliaryData (SophieLedgerEra era))

→ TxScriptValidity era

Mark script as expected to pass or fail validation

→ TxBody era

Bundled Patterns

pattern TxBody ∷ TxBodyContent ViewTx era → TxBody era

makeTransactionBody ∷ ∀ era. IsBccEra era ⇒ TxBodyContent BuildTx era → Either TxBodyError (TxBody era)

data TxBodyContent build era

Constructors

TxBodyContent

Fields

txIns ∷ TxIns build era

txInsCollateral ∷ TxInsCollateral era

txOuts ∷ [TxOut era]

txFee ∷ TxFee era

txValidityRange ∷ (TxValidityLowerBound era, TxValidityUpperBound era)

txMetadata ∷ TxMetadataInEra era

txAuxScripts ∷ TxAuxScripts era

txExtraScriptData ∷ BuildTxWith build (TxExtraScriptData era)

txExtraKeyWits ∷ TxExtraKeyWitnesses era

txProtocolParams ∷ BuildTxWith build (Maybe ProtocolParameters)

txWithdrawals ∷ TxWithdrawals build era

txCertificates ∷ TxCertificates build era

txUpdateProposal ∷ TxUpdateProposal era

txMintValue ∷ TxMintValue build era

txScriptValidity ∷ TxScriptValidity era

data TxBodyError

Constructors

TxBodyEmptyTxIns

TxBodyEmptyTxInsCollateral

TxBodyEmptyTxOuts

TxBodyOutputNegative Quantity TxOutInAnyEra

TxBodyOutputOverflow Quantity TxOutInAnyEra

TxBodyMetadataError [(Word64, TxMetadataRangeError)]

TxBodyMintBccError

TxBodyMissingProtocolParams

data TxBodyScriptData era where

Constructors

TxBodyNoScriptData ∷ TxBodyScriptData era

TxBodyScriptData ∷ ScriptDataSupportedInEra era → TxDats (SophieLedgerEra era) → Redeemers (SophieLedgerEra era) → TxBodyScriptData era

data TxScriptValidity era where

A representation of whether the era supports tx script validity.

The Jen and subsequent eras support script validity.

Constructors

TxScriptValidityNone ∷ TxScriptValidity era

TxScriptValidity ∷ TxScriptValiditySupportedInEra era → ScriptValidity → TxScriptValidity era
Tx script validity is supported in transactions in the Aurum era onwards.

data TxScriptValiditySupportedInEra era where

Constructors

TxScriptValiditySupportedInAurumEra ∷ TxScriptValiditySupportedInEra AurumEra

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.

scriptValidityToIsValid ∷ ScriptValidity → IsValid

isValidToScriptValidity ∷ IsValid → ScriptValidity

scriptValidityToTxScriptValidity ∷ SophieBasedEra era → ScriptValidity → TxScriptValidity era

txScriptValidityToIsValid ∷ TxScriptValidity era → IsValid

txScriptValidityToScriptValidity ∷ TxScriptValidity era → ScriptValidity

Transaction Ids

newtype TxId

Constructors

TxId (Hash StandardCrypto EraIndependentTxBody)

getTxId ∷ ∀ era. TxBody era → TxId

Calculate the transaction identifier for a TxBody.

getTxIdSophie ∷ Crypto (SophieLedgerEra era) ~ StandardCrypto ⇒ UsesTxBody (SophieLedgerEra era) ⇒ SophieBasedEra era → TxBody (SophieLedgerEra era) → TxId

Transaction inputs

data TxIn

Constructors

TxIn TxId TxIx

newtype TxIx

Constructors

TxIx Word

genesisUTxOPseudoTxIn ∷ NetworkId → Hash GenesisUTxOKey → TxIn

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Sophie initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

Transaction outputs

data TxOut era

Constructors

TxOut (AddressInEra era) (TxOutValue era) (TxOutDatumHash era)

data TxOutValue era where

Constructors

TxOutBccOnly ∷ OnlyBccSupportedInEra era → Entropic → TxOutValue era

TxOutValue ∷ MultiAssetSupportedInEra era → Value → TxOutValue era

entropicToTxOutValue ∷ IsBccEra era ⇒ Entropic → TxOutValue era

prettyRenderTxOut ∷ TxOutInAnyEra → Text

txOutValueToEntropic ∷ TxOutValue era → Entropic

txOutValueToValue ∷ TxOutValue era → Value

serialiseAddressForTxOut ∷ AddressInEra era → Text

data TxOutDatumHash era where

Constructors

TxOutDatumHashNone ∷ TxOutDatumHash era

TxOutDatumHash ∷ ScriptDataSupportedInEra era → Hash ScriptData → TxOutDatumHash era

data TxOutInAnyEra where

Constructors

TxOutInAnyEra ∷ BccEra era → TxOut era → TxOutInAnyEra

txOutInAnyEra ∷ IsBccEra era ⇒ TxOut era → TxOutInAnyEra

Convenience constructor for TxOutInAnyEra

Other transaction body types

data TxInsCollateral era where

Constructors

TxInsCollateralNone ∷ TxInsCollateral era

TxInsCollateral ∷ CollateralSupportedInEra era → [TxIn] → TxInsCollateral era

data TxFee era where

Constructors

TxFeeImplicit ∷ TxFeesImplicitInEra era → TxFee era

TxFeeExplicit ∷ TxFeesExplicitInEra era → Entropic → TxFee era

data TxValidityLowerBound era where

Constructors

TxValidityNoLowerBound ∷ TxValidityLowerBound era

TxValidityLowerBound ∷ ValidityLowerBoundSupportedInEra era → SlotNo → TxValidityLowerBound era

data TxValidityUpperBound era where

This was formerly known as the TTL.

Constructors

TxValidityNoUpperBound ∷ ValidityNoUpperBoundSupportedInEra era → TxValidityUpperBound era

TxValidityUpperBound ∷ ValidityUpperBoundSupportedInEra era → SlotNo → TxValidityUpperBound era

data TxMetadataInEra era where

Constructors

TxMetadataNone ∷ TxMetadataInEra era

TxMetadataInEra ∷ TxMetadataSupportedInEra era → TxMetadata → TxMetadataInEra era

data TxAuxScripts era where

Constructors

TxAuxScriptsNone ∷ TxAuxScripts era

TxAuxScripts ∷ AuxScriptsSupportedInEra era → [ScriptInEra era] → TxAuxScripts era

data TxExtraScriptData era where

Constructors

TxExtraScriptDataNone ∷ TxExtraScriptData era

TxExtraScriptData ∷ ScriptDataSupportedInEra era → [ScriptData] → TxExtraScriptData era

data TxExtraKeyWitnesses era where

Constructors

TxExtraKeyWitnessesNone ∷ TxExtraKeyWitnesses era

TxExtraKeyWitnesses ∷ TxExtraKeyWitnessesSupportedInEra era → [Hash PaymentKey] → TxExtraKeyWitnesses era

data TxWithdrawals build era where

Constructors

TxWithdrawalsNone ∷ TxWithdrawals build era

TxWithdrawals ∷ WithdrawalsSupportedInEra era → [(StakeAddress, Entropic, BuildTxWith build (Witness WitCtxStake era))] → TxWithdrawals build era

data TxCertificates build era where

Constructors

TxCertificatesNone ∷ TxCertificates build era

TxCertificates ∷ CertificatesSupportedInEra era → [Certificate] → BuildTxWith build (Map StakeCredential (Witness WitCtxStake era)) → TxCertificates build era

data TxUpdateProposal era where

Constructors

TxUpdateProposalNone ∷ TxUpdateProposal era

TxUpdateProposal ∷ UpdateProposalSupportedInEra era → UpdateProposal → TxUpdateProposal era

data TxMintValue build era where

Constructors

TxMintNone ∷ TxMintValue build era

TxMintValue ∷ MultiAssetSupportedInEra era → Value → BuildTxWith build (Map PolicyId (ScriptWitness WitCtxMint era)) → TxMintValue build era

Building vs viewing transactions

data BuildTxWith build a where

Constructors

ViewTx ∷ BuildTxWith ViewTx a

BuildTxWith ∷ a → BuildTxWith BuildTx a

data BuildTx

data ViewTx

Era-dependent transaction body features

data CollateralSupportedInEra era where

A representation of whether the era supports transactions with inputs used only for collateral for script fees.

The Aurum and subsequent eras support collateral inputs.

Constructors

CollateralInAurumEra ∷ CollateralSupportedInEra AurumEra

data MultiAssetSupportedInEra era where

A representation of whether the era supports multi-asset transactions.

The Jen and subsequent eras support multi-asset transactions.

The negation of this is OnlyBccSupportedInEra.

Constructors

MultiAssetInJenEra ∷ MultiAssetSupportedInEra JenEra
Multi-asset transactions are supported in the Jen era.

MultiAssetInAurumEra ∷ MultiAssetSupportedInEra AurumEra
Multi-asset transactions are supported in the Aurum era.

data OnlyBccSupportedInEra era where

A representation of whether the era supports only bcc transactions.

Prior to the Jen era only bcc transactions are supported. Multi-assets are supported from the Jen era onwards.

This is the negation of MultiAssetSupportedInEra. It exists since we need evidence to be positive.

Constructors

BccOnlyInColeEra ∷ OnlyBccSupportedInEra ColeEra

BccOnlyInSophieEra ∷ OnlyBccSupportedInEra SophieEra

BccOnlyInEvieEra ∷ OnlyBccSupportedInEra EvieEra

data TxFeesExplicitInEra era where

A representation of whether the era requires explicitly specified fees in transactions.

The Cole era tx fees are implicit (as the difference bettween the sum of outputs and sum of inputs), but all later eras the fees are specified in the transaction explicitly.

Constructors

TxFeesExplicitInSophieEra ∷ TxFeesExplicitInEra SophieEra

TxFeesExplicitInEvieEra ∷ TxFeesExplicitInEra EvieEra

TxFeesExplicitInJenEra ∷ TxFeesExplicitInEra JenEra

TxFeesExplicitInAurumEra ∷ TxFeesExplicitInEra AurumEra

data TxFeesImplicitInEra era where

A representation of whether the era requires implicitly specified fees in transactions.

This is the negation of TxFeesExplicitInEra.

Constructors

TxFeesImplicitInColeEra ∷ TxFeesImplicitInEra ColeEra

data ValidityUpperBoundSupportedInEra era where

A representation of whether the era supports transactions with an upper bound on the range of slots in which they are valid.

The Sophie and subsequent eras support an upper bound on the validity range. In the Sophie era specifically it is actually required. It is optional in later eras.

Constructors

ValidityUpperBoundInSophieEra ∷ ValidityUpperBoundSupportedInEra SophieEra

ValidityUpperBoundInEvieEra ∷ ValidityUpperBoundSupportedInEra EvieEra

ValidityUpperBoundInJenEra ∷ ValidityUpperBoundSupportedInEra JenEra

ValidityUpperBoundInAurumEra ∷ ValidityUpperBoundSupportedInEra AurumEra

data ValidityNoUpperBoundSupportedInEra era where

A representation of whether the era supports transactions having no upper bound on the range of slots in which they are valid.

Note that the SophieEra does not support omitting a validity upper bound. It was introduced as a required field in Sophie and then made optional in Evie and subsequent eras.

The Cole era supports this by virtue of the fact that it does not support validity ranges at all.

Constructors

ValidityNoUpperBoundInColeEra ∷ ValidityNoUpperBoundSupportedInEra ColeEra

ValidityNoUpperBoundInEvieEra ∷ ValidityNoUpperBoundSupportedInEra EvieEra

ValidityNoUpperBoundInJenEra ∷ ValidityNoUpperBoundSupportedInEra JenEra

ValidityNoUpperBoundInAurumEra ∷ ValidityNoUpperBoundSupportedInEra AurumEra

data ValidityLowerBoundSupportedInEra era where

A representation of whether the era supports transactions with a lower bound on the range of slots in which they are valid.

The Evie and subsequent eras support an optional lower bound on the validity range. No equivalent of ValidityNoUpperBoundSupportedInEra is needed since all eras support having no lower bound.

Constructors

ValidityLowerBoundInEvieEra ∷ ValidityLowerBoundSupportedInEra EvieEra

ValidityLowerBoundInJenEra ∷ ValidityLowerBoundSupportedInEra JenEra

ValidityLowerBoundInAurumEra ∷ ValidityLowerBoundSupportedInEra AurumEra

data TxMetadataSupportedInEra era where

A representation of whether the era supports transaction metadata.

Transaction metadata is supported from the Sophie era onwards.

Constructors

TxMetadataInSophieEra ∷ TxMetadataSupportedInEra SophieEra

TxMetadataInEvieEra ∷ TxMetadataSupportedInEra EvieEra

TxMetadataInJenEra ∷ TxMetadataSupportedInEra JenEra

TxMetadataInAurumEra ∷ TxMetadataSupportedInEra AurumEra

data AuxScriptsSupportedInEra era where

A representation of whether the era supports auxiliary scripts in transactions.

Auxiliary scripts are supported from the Evie era onwards.

Constructors

AuxScriptsInEvieEra ∷ AuxScriptsSupportedInEra EvieEra

AuxScriptsInJenEra ∷ AuxScriptsSupportedInEra JenEra

AuxScriptsInAurumEra ∷ AuxScriptsSupportedInEra AurumEra

data TxExtraKeyWitnessesSupportedInEra era where

A representation of whether the era supports transactions that specify in the body that they need extra key witnesses, and where this fact is visible to scripts.

Extra key witnesses visible to scripts are supported from the Aurum era onwards.

Constructors

ExtraKeyWitnessesInAurumEra ∷ TxExtraKeyWitnessesSupportedInEra AurumEra

data ScriptDataSupportedInEra era where

A representation of whether the era supports multi-asset transactions.

The Jen and subsequent eras support multi-asset transactions.

The negation of this is OnlyBccSupportedInEra.

Constructors

ScriptDataInAurumEra ∷ ScriptDataSupportedInEra AurumEra
Script data is supported in transactions in the Aurum era.

data WithdrawalsSupportedInEra era where

A representation of whether the era supports withdrawals from reward accounts.

The Sophie and subsequent eras support stake addresses, their associated reward accounts and support for withdrawals from them.

Constructors

WithdrawalsInSophieEra ∷ WithdrawalsSupportedInEra SophieEra

WithdrawalsInEvieEra ∷ WithdrawalsSupportedInEra EvieEra

WithdrawalsInJenEra ∷ WithdrawalsSupportedInEra JenEra

WithdrawalsInAurumEra ∷ WithdrawalsSupportedInEra AurumEra

data CertificatesSupportedInEra era where

A representation of whether the era supports Certificates embedded in transactions.

The Sophie and subsequent eras support such certificates.

Constructors

CertificatesInSophieEra ∷ CertificatesSupportedInEra SophieEra

CertificatesInEvieEra ∷ CertificatesSupportedInEra EvieEra

CertificatesInJenEra ∷ CertificatesSupportedInEra JenEra

CertificatesInAurumEra ∷ CertificatesSupportedInEra AurumEra

data UpdateProposalSupportedInEra era where

A representation of whether the era supports UpdateProposals embedded in transactions.

The Sophie and subsequent eras support such update proposals. They Cole era has a notion of an update proposal, but it is a standalone chain object and not embedded in a transaction.

Constructors

UpdateProposalInSophieEra ∷ UpdateProposalSupportedInEra SophieEra

UpdateProposalInEvieEra ∷ UpdateProposalSupportedInEra EvieEra

UpdateProposalInJenEra ∷ UpdateProposalSupportedInEra JenEra

UpdateProposalInAurumEra ∷ UpdateProposalSupportedInEra AurumEra

Feature availability functions

collateralSupportedInEra ∷ BccEra era → Maybe (CollateralSupportedInEra era)

multiAssetSupportedInEra ∷ BccEra era → Either (OnlyBccSupportedInEra era) (MultiAssetSupportedInEra era)

txFeesExplicitInEra ∷ BccEra era → Either (TxFeesImplicitInEra era) (TxFeesExplicitInEra era)

validityUpperBoundSupportedInEra ∷ BccEra era → Maybe (ValidityUpperBoundSupportedInEra era)

validityNoUpperBoundSupportedInEra ∷ BccEra era → Maybe (ValidityNoUpperBoundSupportedInEra era)

validityLowerBoundSupportedInEra ∷ BccEra era → Maybe (ValidityLowerBoundSupportedInEra era)

txMetadataSupportedInEra ∷ BccEra era → Maybe (TxMetadataSupportedInEra era)

auxScriptsSupportedInEra ∷ BccEra era → Maybe (AuxScriptsSupportedInEra era)

extraKeyWitnessesSupportedInEra ∷ BccEra era → Maybe (TxExtraKeyWitnessesSupportedInEra era)

scriptDataSupportedInEra ∷ BccEra era → Maybe (ScriptDataSupportedInEra era)

withdrawalsSupportedInEra ∷ BccEra era → Maybe (WithdrawalsSupportedInEra era)

certificatesSupportedInEra ∷ BccEra era → Maybe (CertificatesSupportedInEra era)

updateProposalSupportedInEra ∷ BccEra era → Maybe (UpdateProposalSupportedInEra era)

txScriptValiditySupportedInSophieBasedEra ∷ SophieBasedEra era → Maybe (TxScriptValiditySupportedInEra era)

txScriptValiditySupportedInBccEra ∷ BccEra era → Maybe (TxScriptValiditySupportedInEra era)

Inspecting ScriptWitnesses

data AnyScriptWitness era where

A ScriptWitness in any WitCtx. This lets us handle heterogeneous collections of script witnesses from multiple contexts.

Constructors

AnyScriptWitness ∷ ScriptWitness witctx era → AnyScriptWitness era

data ScriptWitnessIndex

Identify the location of a ScriptWitness within the context of a TxBody. These are indexes of the objects within the transaction that need or can use script witnesses: inputs, minted assets, withdrawals and certificates. These are simple numeric indices, enumerated from zero. Thus the indices are not stable if the transaction body is modified.

Constructors

ScriptWitnessIndexTxIn !Word
The n'th transaction input, in the order of the TxIds.

ScriptWitnessIndexMint !Word
The n'th minting PolicyId, in the order of the PolicyIds.

ScriptWitnessIndexCertificate !Word
The n'th certificate, in the list order of the certificates.

ScriptWitnessIndexWithdrawal !Word
The n'th withdrawal, in the order of the StakeAddresss.

renderScriptWitnessIndex ∷ ScriptWitnessIndex → String

collectTxBodyScriptWitnesses ∷ ∀ era. TxBodyContent BuildTx era → [(ScriptWitnessIndex, AnyScriptWitness era)]

mapTxScriptWitnesses ∷ ∀ era. (∀ witctx. ScriptWitnessIndex → ScriptWitness witctx era → ScriptWitness witctx era) → TxBodyContent BuildTx era → TxBodyContent BuildTx era

Internal conversion functions & types

toSophieTxId ∷ TxId → TxId StandardCrypto

toSophieTxIn ∷ TxIn → TxIn StandardCrypto

toSophieTxOut ∷ ∀ era ledgerera. SophieLedgerEra era ~ ledgerera ⇒ SophieBasedEra era → TxOut era → TxOut ledgerera

fromSophieTxId ∷ TxId StandardCrypto → TxId

fromSophieTxIn ∷ TxIn StandardCrypto → TxIn

fromSophieTxOut ∷ SophieLedgerEra era ~ ledgerera ⇒ SophieBasedEra era → TxOut ledgerera → TxOut era

toAurumRdmrPtr ∷ ScriptWitnessIndex → RdmrPtr

fromAurumRdmrPtr ∷ RdmrPtr → ScriptWitnessIndex

fromColeTxIn ∷ TxIn → TxIn

fromLedgerTxOuts ∷ SophieBasedEra era → TxBody (SophieLedgerEra era) → [TxOut era]

renderTxIn ∷ TxIn → Text

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.