Bcc.Api.Value

Description

Currency values

Documentation

newtype Entropic

Constructors

Entropic Integer

Multi-asset values

newtype Quantity

Constructors

Quantity Integer

newtype PolicyId

Constructors

PolicyId ScriptHash

scriptPolicyId ∷ Script lang → PolicyId

newtype AssetName

Constructors

AssetName ByteString

data AssetId

Constructors

BccAssetId

AssetId !PolicyId !AssetName

data Value

selectAsset ∷ Value → AssetId → Quantity

valueFromList ∷ [(AssetId, Quantity)] → Value

valueToList ∷ Value → [(AssetId, Quantity)]

filterValue ∷ (AssetId → Bool) → Value → Value

negateValue ∷ Value → Value

This lets you write a - b as a <> negateValue b.

calcMinimumDeposit ∷ Value → Entropic → Entropic

Calculate cost of making a UTxO entry for a given Value and mininimum UTxO value derived from the ProtocolParameters

Bcc / Entropic specifically

quantityToEntropic ∷ Quantity → Entropic

entropicToQuantity ∷ Entropic → Quantity

selectEntropic ∷ Value → Entropic

entropicToValue ∷ Entropic → Value

valueToEntropic ∷ Value → Maybe Entropic

Check if the Value consists of only Entropic and no other assets, and if so then return the Entropic.

See also selectEntropic to select the Entropic quantity from the Value, ignoring other assets.

Alternative nested representation

newtype ValueNestedRep

An alternative nested representation for Value that groups assets that share a PolicyId.

Constructors

ValueNestedRep [ValueNestedBundle]

data ValueNestedBundle

A bundle within a ValueNestedRep for a single PolicyId, or for the special case of bcc.

Constructors

ValueNestedBundleBcc Quantity

ValueNestedBundle PolicyId (Map AssetName Quantity)

valueToNestedRep ∷ Value → ValueNestedRep

valueFromNestedRep ∷ ValueNestedRep → Value

Rendering

renderValue ∷ Value → Text

Render a textual representation of a Value.

renderValuePretty ∷ Value → Text

Render a "prettified" textual representation of a Value.

Internal conversion functions

toColeEntropic ∷ Entropic → Maybe Entropic

fromColeEntropic ∷ Entropic → Entropic

toSophieEntropic ∷ Entropic → Coin

fromSophieEntropic ∷ Coin → Entropic

fromSophieDeltaEntropic ∷ DeltaCoin → Entropic

toJenValue ∷ Value → Value StandardCrypto

fromJenValue ∷ Value StandardCrypto → Value

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.